Привел в порядок репозиторий, добавил грэдл, обновил.gitignore, добавил папку resources, поменял картинку, убрал пробелы в FirstJUnitTest
