+++ .gitignore
@@ -10,6 +10,9 @@
 # Mobile Tools for Java (J2ME)
 .mtj.tmp/
 
+# Properties files
+*.properties
+
 # Package Files #
 *.jar
 *.war
@@ -18,6 +21,13 @@
 *.zip
 *.tar.gz
 *.rar
+*.iml
 
-# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
-hs_err_pid*
+# Project gen files
+/.gradle/
+/.idea/
+/build/
+/src/main/
+/gradle/
+/gradlew
+/gradlew.bat
